requirements.txt
kafka-python
pyspark
streamlit
requests
pandas
plotly
python-dotenv

#this are requirements that is needed

